further cleanup + address PR comments
